Fix logical readers reference
