+++ fastavro/_logical_readers_py.pyi
@@ -1,0 +1,3 @@
+from typing import Any, Dict, Callable
+
+LOGICAL_READERS: Dict[str, Callable[[Any, Dict], Any]]
